backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2==2.9.7
redis==4.5.0
spacy==3.5.0
numpy==1.23.5
thinc==8.1.7
python-dotenv==1.0.0
